refactor: use fs promises functions in all tasks
