+++ src/fs/copy.js
@@ -1,5 +1,5 @@
 import * as url from 'url';
-import fs from 'fs';
+import fs from 'fs/promises';
 import path from 'path';
 
 const __filename = url.fileURLToPath(import.meta.url);
@@ -8,22 +8,26 @@ const dirToCopy = path.join(__dirname, 'files');
 const copyDest = path.join(__dirname, 'files_copy');
 
 const copy = async () => {
-  fs.access(dirToCopy, fs.F_OK, err => {
-    if (err) {
+  try {
+    await fs.access(dirToCopy, fs.F_OK);
+  } catch {
+    throw new Error('FS operation failed');
+  }
+
+  try {
+    await fs.access(copyDest, fs.F_OK);
+    throw new Error('FS operation failed');
+  } catch (err) {
+    if (err.message === 'FS operation failed') {
       throw new Error('FS operation failed');
     }
-  });
+  }
 
-  fs.access(copyDest, fs.F_OK, err => {
-    if (err) return;
-    throw new Error('FS operation failed');
-  });
-
-  fs.cp(dirToCopy, copyDest, { recursive: true }, err => {
-    if (err) {
-      console.error(err);
-    }
-  });
+  try {
+    await fs.cp(dirToCopy, copyDest, { recursive: true });
+  } catch (err) {
+    console.error(err);
+  }
 };
 
 await copy();
