+++ src/fs/create.js
@@ -1,5 +1,5 @@
 import * as url from 'url';
-import fs from 'fs';
+import fs from 'fs/promises';
 import path from 'path';
 
 const __filename = url.fileURLToPath(import.meta.url);
@@ -7,17 +7,20 @@ const __dirname = path.dirname(__filename);
 const filePath = path.join(__dirname, 'files', 'fresh.txt');
 
 const create = async () => {
-  fs.access(filePath, fs.F_OK, err => {
-    if (err) return;
-
-    //file exists
+  try {
+    await fs.access(filePath, fs.F_OK);
     throw new Error('FS operation failed');
-  });
-
-  fs.appendFile(filePath, 'I am fresh and young', function (err) {
-    if (err) throw new Error('FS operation failed');
+  } catch (err) {
+    if (err.message === 'FS operation failed') {
+      throw new Error('FS operation failed');
+    }
+  }
+  try {
+    await fs.appendFile(filePath, 'I am fresh and young');
     console.log('Saved!');
-  });
+  } catch (err) {
+    console.error(err);
+  }
 };
 
 await create();
